Refactoring to provided extension points for Conduit Worker.
This allows us to treat the Backbone.Conduit data management components as an extensions,
as well as any developer-provided sorting methods the same way.
