+++ gulp-setup/tasks/dist-worker.js
@@ -8,10 +8,20 @@ var gulp = require('gulp');
 var webpack = require('gulp-webpack');
 
 gulp.task('dist:worker', false, ['bower'], function () {
-    return gulp.src('src/worker/worker.js')
+    // Produce the core worker
+    gulp.src('src/worker/index.js')
         .pipe(webpack({
             output: {
                 filename: 'backbone.conduit-worker.js'
+            }
+        }))
+        .pipe(gulp.dest('dist'));
+
+    // Produce the dataManagement component
+    gulp.src('src/worker/dataManagement/index.js')
+        .pipe(webpack({
+            output: {
+                filename: 'conduit.worker.dataManagement.js'
             }
         }))
         .pipe(gulp.dest('dist'));
