+++ spec/worker/dataManagement/dataUtils.spec.js
@@ -6,30 +6,33 @@
 
 var _ = require('underscore');
 
-var dataUtils = require('./../../src/worker/dataUtils');
+var mockConduitWorker = require('../mockConduitWorker');
+
+var dataUtils = require('./../../../src/worker/dataManagement/dataUtils');
 
 describe('The worker/dataUtils module', function() {
     var context, data;
 
     beforeEach(function() {
-        context = {};
+        mockConduitWorker.reset();
+        context = mockConduitWorker.get();
         data = this.getSampleData();
     });
 
     it('initializes the context to have a "data" property', function() {
-        dataUtils.initStore(context);
+        dataUtils.initStore();
         expect(context.data).to.be.an('array');
     });
 
     it('will not recreate the data store by default', function() {
         context.data = [ { foo: "bar" }];
-        dataUtils.initStore(context);
+        dataUtils.initStore();
         expect(context.data).to.have.length(1);
     });
 
     it('will recreate the data store if explicitly requested', function() {
         context.data = [ { foo: "bar" } ];
-        dataUtils.initStore(context, { reset: true });
+        dataUtils.initStore({ reset: true });
         expect(context.data).to.have.length(0);
     });
 
@@ -52,8 +55,8 @@ describe('The worker/dataUtils module', function() {
 
     describe('after initializing and adding data', function() {
         beforeEach(function() {
-            dataUtils.initStore(context);
-            dataUtils.addTo(context, this.getSampleData());
+            dataUtils.initStore();
+            dataUtils.addTo(this.getSampleData());
         });
 
         it('contains the fresh data', function() {
@@ -62,7 +65,7 @@ describe('The worker/dataUtils module', function() {
         });
 
         it('can add data to existing', function() {
-            dataUtils.addTo(context, [
+            dataUtils.addTo([
                 { id: 10, name: "ten", first: 6, second: 4 }
             ]);
 
@@ -71,7 +74,7 @@ describe('The worker/dataUtils module', function() {
         });
 
         it('can merge data with existing', function() {
-            dataUtils.addTo(context, [
+            dataUtils.addTo([
                 { id: 3, name: "THREE" }
             ]);
             expect(context.data).to.have.length(3);
@@ -86,7 +89,7 @@ describe('The worker/dataUtils module', function() {
         });
 
         it('sets "_dataIndex" property to added data properly', function() {
-            dataUtils.addTo(context, [
+            dataUtils.addTo([
                 { id: 10, name: "ten", first: 6, second: 4 }
             ]);
             for (var i = 0; i < context.data.length; i++) {
@@ -96,7 +99,7 @@ describe('The worker/dataUtils module', function() {
         });
 
         it('does not disturb the "_dataIndex" property of merged data', function() {
-            dataUtils.addTo(context, [
+            dataUtils.addTo([
                 { id: 3, name: "THREE" }
             ]);
             for (var i = 0; i < context.data.length; i++) {
