+++ src/worker/dataManagement/index.js
@@ -1,0 +1,19 @@
+'use strict';
+
+/**
+ * This provides a Conduit.Worker component for managing data on the worker.
+ */
+
+if (typeof ConduitWorker !== 'undefined') {
+    // Register our component
+    ConduitWorker.registerComponent({
+        name: 'dataManagement',
+
+        methods: [
+            require('./setData'),
+            require('./mergeData'),
+            require('./prepare'),
+            require('./sortBy')
+        ]
+    });
+}
